Fix circular imports with shared types module
- Create nclone/graph/subgoal_types.py with shared data classes
- Move Subgoal, SubgoalPlan, CompletionStrategyInfo to shared module
- Remove duplicate CompletionStrategyInfo from reachability_types.py
- Update all imports to use shared types module
- Remove level_completion_analyzer.py (unused duplicate functionality)
- Fix lazy imports in game_mechanics.py

This resolves the circular import chain:
SubgoalPlanner ↔ reachability modules ↔ game_mechanics ↔ SubgoalPlanner

All systems now import cleanly and work correctly.
